WIP: Don't run tests if the profile is not the same as intended.
