+++ Assets/Digital Painting/Scripts/Utility/StringExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b475f71ee0bc3374f9d4c7fc1e7e8c19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
