+++ Assets/Digital Painting/Scripts/Validation/ValidationResultsCollection.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 203b9e0f06948af44bce38eb0e554557
+guid: c28525f58d5ee3d489e368a79f6b1147
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
