+++ Assets/Digital Painting/Scripts/Validation/ValidationTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 069b552bb371bfa4ca92f1d171c35df6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
